requirements.txt
alembic==1.13.1
Babel==2.14.0
bcrypt==4.1.3
bleach==6.1.0
blinker==1.7.0
chardet==5.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
flask-babel==4.0.0
Flask-DebugToolbar==0.14.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
packaging==24.0
passlib==1.7.4
pillow==10.3.0
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
reportlab==4.2.0
six==1.16.0
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
typing_extensions==4.11.0
webencodings==0.5.1
Werkzeug==3.0.2
WTForms==3.1.2
